restore page rework

+++ web/concrete/tools/pages/delete.php
@@ -16,9 +16,16 @@ if ($_POST['task'] == 'delete_pages') {
 			$cp = new Permissions($c);
 			$children = $c->getNumChildren();
 			if ($children == 0 || $cp->canDeletePage()) {
-				$c->markPendingAction('DELETE');
-				if ($cp->canApprovePageVersions()) {
-					$c->delete();
+				if ($cp->canApprovePageVersions()) { 
+					if (ENABLE_TRASH_CAN) {
+						$c->moveToTrash();
+					} else {
+						$c->delete();
+					}
+				} else { 
+					$pkr = new DeletePagePageWorkflowRequest();
+					$pkr->setRequestedPage($c);
+					$pkr->trigger();
 				}
 			} else {
 				$json['error'] = t('Unable to delete one or more pages.');
